+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added fallback on DNS failure
 * Update to latest srvclient
 * Switch to RPCLib's result
+* Added preprocess function
 
 ### 0.2.0 ###
 * RPCResult can be treated as a promise